Update demo.py to automatically download model
Keras uses get_file() to do this when loading the [inception models](https://github.com/keras-team/keras/blob/master/keras/applications/inception_v3.py).

It checks if this model is already downloaded, if not, then it automatically downloads it to your hard drive.  It knows where to look the next time you load it.